+++ src/components/DropDown/test/DropDown.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import Paper from 'material-ui/Paper';
 import Input, { InputLabel } from 'material-ui/Input';
-import {DropDown,DropdownItem} from '../DropDown';
+import { DropDown, DropdownItem } from '../../DropDown';
 const stories = storiesOf('DropDown', module);
 
 const onChange =(event) => {
@@ -18,16 +18,16 @@ const style={
 };
 const multiItems=['Oliver Hansen',
     'Van Henry',
-    'April Tucker',
-    'Ralph Hubbard',
-    'Omar Alexander',
-    'Carlos Abbott',
-    'Miriam Wagner',
-    'Bradley Wilkerson',
-    'Virginia Andrews',
-    'Kelly Snyder',];
+	'April Tucker',
+	'Ralph Hubbard',
+	'Omar Alexander',
+	'Carlos Abbott',
+	'Miriam Wagner',
+	'Bradley Wilkerson',
+	'Virginia Andrews',
+	'Kelly Snyder',];
 stories.add('DropDown', () => {
-    return(
+	return(
     <Paper style={style.root}>
         <DropDown onValueChange={onChange} value={'xyz'}  style={style.dropdown}>
           <DropdownItem value="xyz">Item 1 </DropdownItem>
@@ -39,7 +39,7 @@ stories.add('DropDown', () => {
     </Paper>);}
 );
 stories.add('DropDown with autoWidth', () => {
-    return(
+	return(
         <Paper style={style.root}>
             <DropDown onValueChange={onChange} value={'xyz'} autoWidth={true}  style={style.dropdown}>
                 <DropdownItem value="xyz">Item 1 </DropdownItem>
@@ -50,9 +50,9 @@ stories.add('DropDown with autoWidth', () => {
             </DropDown>
         </Paper>);}
 );stories.add('DropDown with Override styles', () => {
-    return(
+	return(
         <Paper style={style.root}>
-            <DropDown onValueChange={onChange} value={'xyz'} style={{width:600, backgroundColor:'green'}} itemStyle={{backgroundColor:'black'}}>
+            <DropDown onValueChange={onChange} value={'xyz'} style={{ width:600, backgroundColor:'green' }} itemStyle={{ backgroundColor:'black' }}>
                 <DropdownItem value="xyz">Item 1 </DropdownItem>
                 <DropdownItem value="abc">Item 2 </DropdownItem>
                 <DropdownItem value="abc">Item 3 </DropdownItem>
@@ -61,7 +61,7 @@ stories.add('DropDown with autoWidth', () => {
             </DropDown>
         </Paper>);}
 );stories.add('DropDown multiple select', () => {
-    return(
+	return(
         <Paper style={style.root}>
             <DropDown onValueChange={onChange} value={multiItems} autoWidth={true} multiple style={style.dropdown} >
                 <DropdownItem value="xyz">Item 1 </DropdownItem>
@@ -72,7 +72,7 @@ stories.add('DropDown with autoWidth', () => {
             </DropDown>
         </Paper>);}
 );stories.add('DropDown with disabled', () => {
-    return(
+	return(
         <Paper style={style.root}>
             <DropDown onValueChange={onChange} value={'xyz'}  input={<Input id="name-native-disabled" />} style={style.dropdown} displayEmpty>
                 <DropdownItem value="xyz">Item 1 </DropdownItem>
