+++ src/components/List/tests/ListItemRightButton.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import  { ListItemRightButton } from '../../List';
-import Button from '../../button';
+import Button from '../../Button';
 import BR from '@blueeast/bluerain-os';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
@@ -11,7 +11,7 @@ const Text = BR.Components.get('Text');
 const Image = BR.Components.get('Image');
 
 const label = text('label as child', 'with knobs');
-storiesOf('ListItemRightButton', module).add('Card with style and raised props', () =>(
+storiesOf('ListItemRightButton', module).add('Card with style and raised props', () => (
 <ListItemRightButton
     style={{ backgroundColor:'grey' }}
 >
