+++ src/components/List/tests/ListItemText.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import  { ListItemText } from '../../List';
-import Button from '../../button';
+import Button from '../../Button';
 import BR from '@blueeast/bluerain-os';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
