+++ src/index.tsx
@@ -2,14 +2,14 @@ import React from 'react';
 import { Plugin, BlueRainType } from '@blueeast/bluerain-os';
 
 import TextInput from './components/TextInput';
-import Chip from './components/chip';
-import Badge from './components/badge';
-import Avatar from './components/avatar';
+import Chip from './components/Chip';
+import Badge from './components/Badge';
+import Avatar from './components/Avatar';
 import Switch from './components/Switch';
-import { DropDown, DropdownItem } from './components/DropDown/DropDown';
+import { DropDown, DropdownItem } from './components/DropDown';
 import List, { ListItem, ListHeader, ListItemAvatar, ListItemIcon, ListItemRightButton, ListItemText } from './components/List';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import Button from './components/button';
+import Button from './components/Button';
 import IconComponent from './components/Icon';
 import Card, { CardHeader, CardBody, CardFooter, CardMedia } from './components/Card';
 import Breadcrumb from './components/Breadcrumb';
