+++ test/Badge.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { create } from 'react-test-renderer';
-import Badge from '../src/components/badge/badge';
+import Badge from '../src/components/Badge';
 
 test('Badge component with required fields', () => {
   const badge = create(
@@ -12,8 +12,8 @@ test('Badge component with required fields', () => {
 
 test('Badge component with style', () => {
   const badge = create(
-      <Badge badgeContent={'123'} style={{'background-color':'green'}}></Badge>
+      <Badge badgeContent={'123'} style={{ 'background-color':'green' }}></Badge>
   ).toJSON();
-  expect(badge.props.style).toEqual({'background-color':'green'});
-  expect(badge).toMatchSnapshot();
+  expect(badge.props.style).toEqual({ 'background-color':'green' });
+	expect(badge).toMatchSnapshot();
 });
